+++ src/components/board/components/tasksBlock.js
@@ -1,6 +1,6 @@
 import {Count} from "../../consts";
-import {createTaskEdit} from "../../tasks/task-edit-form/index-edit";
-import {createTasks} from "../../tasks/common/tasks-creation";
+import {createTask} from "../../tasks/index-view";
+import {createTasks} from "../../tasks/components/common/tasks-creation";
 import {generateTasks} from "../../../mock/components/task";
 
 export const tasks = generateTasks(Count.TASKS);
@@ -12,7 +12,7 @@ export const tasks = generateTasks(Count.TASKS);
 export const createBoardTasks = () => {
   return (`
     <div class="board__tasks">
-      ${createTaskEdit(tasks[0])}
+      ${createTask(tasks[0], false)}
       ${createTasks(tasks.slice(1, Count.TASKS_ON_START))}
     </div>
   `);
