+++ src/components/tasks/components/common/buttonsBlock.js
@@ -1,10 +1,11 @@
-export const createButtonsBlock = (task) => {
+export const createButtonsBlock = (isView, task) => {
   const {isArchive, isFavorite} = task;
 
   const archiveButtonInactiveClass = isArchive ? `` : `card__btn--disabled`;
   const favoriteButtonInactiveClass = isFavorite ? `` : `card__btn--disabled`;
 
-  return (`
+  if (isView) {
+    return (`
     <div class="card__control">
       <button type="button" class="card__btn card__btn--edit">
         edit
@@ -22,4 +23,12 @@ export const createButtonsBlock = (task) => {
       </button>
     </div>
   `);
+  } else {
+    return (`
+    <div class="card__status-btns">
+      <button class="card__save" type="submit">save</button>
+      <button class="card__delete" type="button">delete</button>
+    </div>
+  `);
+  }
 };
