+++ src/components/tasks/components/common/descriptionBlock.js
@@ -1,10 +1,18 @@
 /**
  * Создание разметки блока с описанием задачи
  * @param {string} description описание задачи
+ * @param {Boolean} isView
  * @return {string} разметка блока с описанием задачи
  */
-export const createDescription = (description) => {
-  return (`
+export const createDescription = (description, isView) => {
+  if (isView) {
+    return (`
+    <div class="card__textarea-wrap">
+      <p class="card__text">${description}</p>
+    </div>
+  `);
+  } else {
+    return (`
     <div class="card__textarea-wrap">
       <label>
         <textarea
@@ -15,4 +23,5 @@ export const createDescription = (description) => {
       </label>
     </div>
   `);
+  }
 };
