--- src/components/tasks/task-edit-form/components/buttonsBlock.js
@@ -1,12 +1,0 @@
-/**
- * Создание разметкир блока кнопок управления задачей
- * @return {string} разметка блока кнопок управления задачей
- */
-export const createButtonsBlock = () => {
-  return (`
-    <div class="card__status-btns">
-      <button class="card__save" type="submit">save</button>
-      <button class="card__delete" type="button">delete</button>
-    </div>
-  `);
-};
